gdutils/requirements.txt
azure-cognitiveservices-speech
bs4
cachetools
elevenlabs
google-cloud-texttospeech
google-cloud-translate
humanize
ipdb
jinja2
keyboard
llm
openai
pendulum
# for playing mp3
playsound
# for playing mp3
python-vlc
# for playing mp3
pygame

